+++ src/lib/responses.ts
@@ -1,5 +1,5 @@
 import { Message } from "discord.js";
-import { client } from "../bot";
+import { sendMessage } from "./message";
 import { delay } from "./util";
 require("dotenv").config();
 
@@ -20,15 +20,14 @@ export async function respondToMessage(message: Message): Promise<void> {
   responses.forEach(async (pr: PhraseResponder) => {
     pr.phrases.forEach(async (phrase: string) => {
       if (message.content.toLowerCase().includes(phrase)) {
-        await delay(Math.random() * 1000 + 1000);
-        message.channel.sendTyping();
-        await delay(Math.random() * 750 + 750);
-        if (pr.removeAfter)
-          message.reply(pr.response).then(async (msg: Message) => {
-            await delay(pr.removeAfter || 10000);
-            msg.delete();
-          });
-        else message.reply(pr.response);
+        sendMessage(message, pr.response, 1000)
+          .then(async (msg) => {
+            if (msg && pr.removeAfter) {
+              await delay(pr.removeAfter || 10000);
+              msg.delete();
+            }
+          })
+          .catch(console.error);
         return;
       }
     });
